- rename both samples to better reflect their content:   - from "EyeGazeAndBackgroundSegmentation" to "WinRT_ExtendedControlAndMetadataSample" (namespace, app folder, sample folder)   - from  "MediaCaptureSampleExtendedControlsÈ to "CPP-CLI_MediaCaptureSampleExtendedControl" (simply the folder) - add and update readme documents - nit code fix and addition to better support the new DDIs and to help extending the sample
