+++ Samples/ExtendedControlAndMetadata/CPP-CLI_MediaCaptureSampleExtendedControl/App.xaml.h
