+++ Samples/ExtendedControlAndMetadata/CPP-CLI_MediaCaptureSampleExtendedControl/Assets/Square44x44Logo.targetsize-24_altform-unplated.png
