+++ Samples/ExtendedControlAndMetadata/CPP-CLI_MediaCaptureSampleExtendedControl/Assets/Wide310x150Logo.scale-200.png
