+++ Samples/ExtendedControlAndMetadata/CPP-CLI_MediaCaptureSampleExtendedControl/MainPage.xaml.cpp
