+++ Samples/ExtendedControlAndMetadata/CPP-CLI_MediaCaptureSampleExtendedControl/MediaCaptureSampleExtendedControls.vcxproj.filters
