+++ Samples/ExtendedControlAndMetadata/CPP-CLI_MediaCaptureSampleExtendedControl/pch.cpp
