--- Samples/ExtendedControlAndMetadata/EyeGazeAndBackgroundSegmentation/EyeGazeAndBackgroundSegmentation/App.xaml
@@ -1,7 +1,0 @@
-﻿<Application
-    x:Class="EyeGazeAndBackgroundSegmentation.App"
-    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:local="using:EyeGazeAndBackgroundSegmentation">
-
-</Application>
