+++ Samples/ExtendedControlAndMetadata/WinRT_ExtendedControlAndMetadataSample/ExtendedControlAndMetadataSampleApp/App.xaml
@@ -1,0 +1,7 @@
+﻿<Application
+    x:Class="ExtendedControlAndMetadataSampleApp.App"
+    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+    xmlns:local="using:ExtendedControlAndMetadataSampleApp">
+
+</Application>
