+++ Samples/ExtendedControlAndMetadata/WinRT_ExtendedControlAndMetadataSample/ExtendedControlAndMetadataSampleApp/Assets/LockScreenLogo.scale-200.png
