+++ Samples/ExtendedControlAndMetadata/WinRT_ExtendedControlAndMetadataSample/ExtendedControlAndMetadataSampleApp/Assets/SplashScreen.scale-200.png
