+++ Samples/ExtendedControlAndMetadata/WinRT_ExtendedControlAndMetadataSample/NuGet.Config
