+++ Samples/ExtendedControlAndMetadata/WinRT_ExtendedControlAndMetadataSample/example_BACKGROUNDSEGMENTATIONMASK.jpg
